avoid dereferencing a nullpointer
When the model->instances_by_type(m.first) can not find the matching instance, it returns aggregate_of_instance::ptr(), which causes the boost::shared_ptr<IfcEntityList> entities to be a nullpointer. In the if condition by controlling "entities->size() > 0", we end up trying to dereference a nullpointer which causes the programm to crash. To solve this we should controll if the entities is not a nullpointer first. To see the details of the called function "instances_by_type" from IfcFile.h, look: https://github.com/IfcOpenShell/IfcOpenShell/blob/c90c43832432983a53b693dcf0ab6338681cec9b/src/ifcparse/IfcParse.cpp#L2334