avoid dereferencing nullpointers in products_with_faulty_shape
If a product has a NULL shape and the programm fails to heal the shape with  Kernel::heal_products(products_with_faulty_shape), then the functions best_fitting_bbox, obb_to_shape, obb,compound_from_shape_list will be called in the given order, which ends up giving the shape to the TopoDS_Builder::Add(TopoDS_Shape& S, const TopoDS_Shape& C) function as the second argument. But this results in the NULL shape to be dereferenced in the called function, which causes a segmentation fault and terminates the programm.